Refactor:Abstract security context interceptor
