+++ src/main/java/nextstep/auth/AuthConfig.java
@@ -7,8 +7,8 @@ import nextstep.auth.authentication.converter.TokenAuthenticationConverter;
 import nextstep.auth.authentication.interceptor.SessionAuthenticationInterceptor;
 import nextstep.auth.authentication.interceptor.TokenAuthenticationInterceptor;
 import nextstep.auth.authorization.AuthenticationPrincipalArgumentResolver;
-import nextstep.auth.authorization.SessionSecurityContextPersistenceInterceptor;
-import nextstep.auth.authorization.TokenSecurityContextPersistenceInterceptor;
+import nextstep.auth.authorization.interceptor.SessionSecurityContextPersistenceInterceptor;
+import nextstep.auth.authorization.interceptor.TokenSecurityContextPersistenceInterceptor;
 import nextstep.auth.token.JwtTokenProvider;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
