+++ src/main/java/nextstep/auth/authorization/interceptor/SecurityContextInterceptor.java
@@ -1,0 +1,15 @@
+package nextstep.auth.authorization.interceptor;
+
+import nextstep.auth.context.SecurityContextHolder;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+public abstract class SecurityContextInterceptor implements HandlerInterceptor {
+
+    @Override
+    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception {
+        SecurityContextHolder.clearContext();
+    }
+}
