+++ src/main/java/nextstep/auth/authorization/interceptor/SessionSecurityContextPersistenceInterceptor.java
@@ -1,15 +1,15 @@
-package nextstep.auth.authorization;
+package nextstep.auth.authorization.interceptor;
 
+import nextstep.auth.authorization.interceptor.SecurityContextInterceptor;
 import nextstep.auth.context.SecurityContext;
 import nextstep.auth.context.SecurityContextHolder;
-import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import static nextstep.auth.context.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
 
-public class SessionSecurityContextPersistenceInterceptor implements HandlerInterceptor {
+public class SessionSecurityContextPersistenceInterceptor extends SecurityContextInterceptor {
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
         if (SecurityContextHolder.getContext().getAuthentication() != null) {
@@ -21,10 +21,5 @@ public class SessionSecurityContextPersistenceInterceptor implements HandlerInte
             SecurityContextHolder.setContext(securityContext);
         }
         return true;
-    }
-
-    @Override
-    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) {
-        SecurityContextHolder.clearContext();
     }
 }
